refactor: extract virtual scroll script to shared generator
Single source of truth for notebook virtual scrolling:
- New VirtualScrollScriptGenerator.ts in shared/infrastructure/ui/virtualScroll
- DataverseNotebookController now uses generateVirtualScrollScript()
- Added cross-reference comment in VirtualTableRenderer.js

If the core algorithm needs to change, update VirtualScrollScriptGenerator.ts.
VirtualTableRenderer.js has additional panel-specific features (search, sort,
selection) but uses the same spacer row approach.
